Add facemetrix gstreamer element
Add the gstfacemetrix.[ch] files, with the GstElement implementation and
sglclient.[ch] files, with the initial implementation of the SGL client
that is required to access the FaceMetrix server. The code that extracts
the face and converts it to jpeg using the opencv API has not been
tested and so I've added a couple of FIXMEs in the places that should be
reviewed/rewritten.