Update sglclient and adjust the facemetrix plugin code accordingly
The newest version of the facemetrix server added support for
identifying the source of the images, so that the recognition service
won't try to match faces from cameras/video stream that have very
different properties. The sgl client library has been updated to
accomodate these changes, and so has the gstfacemetrix element (we've
added a property -- "source" -- that should be used to identify the
stream).