Enable selection of filename pattern used by the 'save detected faces' feature
Added the 'save_prefix' property that allows one to specify a prefix
that will be used to name the jpeg files with the detected faces. This
is useful when one is using the plugin to select faces that may be used
to train the facemetrix server.

So, for instance, if we know that a video has user 'john' in it, we may
set the property 'save_prefix' as '/home/myusername/training_set/john'
(and the 'source' property as the cameras identifier, say 'camABC'), and
the detected faces will be saved as "john_camABC_<timestamp>.jpg" inside
the directory '/home/myusername/training_set'.

Additionally, I've renamed a couple of properties ('debug' =>
'save_faces', 'sourceid' => 'source') to improve usability a bit.