Added an GstElement based on the 'motempl' opencv example
There's a good change that this plugin is a bit redundant since the
'tracker' plugin is supposed to do something similar. But I needed
something that implemented simple motion tracking using a relatively
fast algorithm for the 'breakout' game/demo. Eventually we might decide
that the functionality provided by this element may be extracted from
the 'tracker' plugin, but there's also a chance that it might make sense
to keep it because, at least in theory, this one should be tweaked for
speed, while the latter should be tweaked for accuracy.

The 'identifier_motion.[ch]' files implements similar functionality, but
it's not as well encapsulated nor does it use the usual glib/gstreamer
conventions. Thus, is might eventually make sense to improve/tweak this
element to the point where we may simply plug it into a pipeline before
the the 'tracker' and/or 'facemetrix' plugins, and drop the
functionality provided by the 'identifier_motion.[ch]' files.

From a tech POV, the plugin is a direct port of the 'motempl' opencv
example. The main addition is that it emits 'element' messages to the
pipeline bus with the average 'deltas' (one for each axis, x and y)
computed across all the detected 'motion components'.