Major refactoring/clean-up of the build system and headers
A lot of the code base of the tracker and facemetrix plugins needs to be
refactored and/or cleaned up. There's problems with inconsistent naming
schemes, code style, mixture of english and portuguese comments and
variable names, just to name a few...

This patch is just a start. I've extracted the common code used by the
two plugins into a 'common' convenience library, and cleaned up the
headers a bit (but no completely; for instance, there is a bunch of
functions that are currently exported for reasons beyond my
understanding. Erickson and Lucas might know the reason though, so they
should have a look before un-exporting them).