Add simplistic font-scale control to the facemetrix element
Previously, the 'facemetrix' element would draw the user id above the
detected face using a font-size that was independent of the video
resolution (and it was a pretty ugly font as well). I've added a
font-scaling parameter to the 'draw face id' function, and a simple test
that defines the font-scaling factor based on the video resolution.

As a bonus, I've fixed some minor styling issues with the draw face id
function...