Split facemetrix plugin into three
So far, all of the opencv plugins have been written in a monolithic
fashion which is less than desirable as it prevents code reuse,
isolation and does not leaverage one of the greatest strengths of the
Gstreamer framework: modularization.

This patch starts to address this issue: I've split the three main
functionalities provided by old 'facemetrix' plugin into three separate
blocks/plugins: background detection, face detection (optionally using
the output of the background detection to speed up the search by
discarding regions where there's no foreground activity), and
face recogtion (which depends on the face detection results).

The original facemetrix plugin provided a fourth functionality which
hasn't been ported yet -- face tracking. The reason is purely pragmatic:
the tracking algorithm doesn't work well enough (yet), so merely using
the detection results yields better looking videos.

Also, it's likely we'll want to eventually use a separate 'drawing'
element after the detection/tracking to actually draw to the screen. The
reason is that by postponing the drawing primitives, the original video
is preserved to enable the addition of other processing element in the
same pipeline.

For the time being, all three elements provide two default properties:
'verbose' and 'display' (both set to 'false' by default). 'verbose'
allows one to view useful debugging messages by setting the GST_DEBUG
variable to INFO (3) level or lower. And 'display' draws the output of
the element's processing into the actual stream (thus, enabling it on
an element that's plugged 'first' in the pipeline might only make sense
for debugging purposes; for instance, if you build a pipeline with
something like
    'src ! bgfgcodebook display=true ! facedetect ! sink'
the output of the face detection algorithm will be 'corrupted' by the
boxes/shadows that the background detection element has drawn).