Add 'objectstracker' element
This element uses the ROIs generated by 'haardetect / haaradjust'. For
each roi, are processed 'surf features' that are stored in a temporary
historical of objects. Throughout the video, these objects are fetched,
updated, deleted and, if new object appears, included. The purpose of
this element is to locate an object ever found by the element
'haardetect / haaradjust', but no more was found by this element.