Enable a few other warning gcc flags; also applied several styling fixes
Enabled a bunch of gcc warning flags to ensure we're more compliant with
the 'C90' standard (which is what most GNU software uses and most folk
recommend to ensure compatibility with MSVC).

Additionally, applied a bunch of styling/indent/whitespace fixes.
There's still several naming conventions that need to be fixes (mostly
regarding the use of camel-casing on function names) but I'll leave
these for another day (or someone else :-)).