Add 'optical flow' (pyramid + lucas kanade) based tracker
This tracker element uses the ROIs supplied by an upstream haaradjust
element and the foreground mask & ROIs (also supplied by a foreground
detection element) to extract features and later track them using the
optical flow algorithm. The exact algorithm used by this version is the
pyramid+lucas-kanade one.

The rule to match objects is very simplistic: we just compare the
percentage of overlap between the haar ROI and the known objects. It's
rather silly but it's the best we could come up with for the time being
-- plus, it works reasonably well for the simpler cases.

Object prunning is done using two simple rules: (1) if the percentage of
matching features between two frames drops below a given threshold or
(2), if a object is tracked for more than a given number of frames
without an event from the upstream haaradjust element -- which is used
as confirmation that the tracked object stil preserves the same
properties.