Add the 'objectdistances' gstreamer element
Split the part of the 'objectsareainteraction' element that handles the
calculating the distance between pairs of static-dynamic or
dynamic-dynamic objects into a separate element. Since our description
of a "TrackedObject" is somewhat specific, we had to came up with a
custom heuristic to calculate the distance: we use the minimum euclidean
distance between each centroid of line segment that composes each
object.

The element emits downstream events with the ids of each object and the
calculated distance between them, and optionally draws a line between
the objects with and overlapping label displaying the calculated
distance.