Add the 'hogdetect' gstreamer element
This element provides object detection using OpenCV's HOG (Histogram of
Oriented Gradients) API. It relies on customizations we've applied to
our local OpenCV tree which add support for calculating confidence
densities during the detection process.

The ROI events are compatible with the 'haardetect' element, thus making
it possible to replace these two elements for some configurations (and
assuming that there are equivalent HOG trained descriptors and HAAR
models; as of this writing, there is only one HOG trained descriptor
available: the 'people detector', which comes bundled and hardcoded
inside the HOG API itself). The confidence density events, however, are
exclusive of this element, are mostly provided to be used as input to
our implementation of Breitenstein's tracking algorithm.